CRUD/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
WTForms==3.0.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
SQLAlchemy==2.0.23
python-dotenv==1.0.0
email-validator==2.1.0.post1